fastmcp_server/pyproject.toml
[project]
name = "fastmcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.8.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-httpx>=0.35.0",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]
